requirements.txt
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
watchdog==4.0.1
pdfkit==1.0.0
lxml==5.3.1
python-docx==1.1.2
google-api-python-client==2.148.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
